Only empty SBTree can be removed.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/local/OSBTree.java
@@ -26,6 +26,7 @@ import com.orientechnologies.common.serialization.types.OBinarySerializer;
 import com.orientechnologies.common.serialization.types.OIntegerSerializer;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.encryption.OEncryption;
+import com.orientechnologies.orient.core.exception.ONonEmptyComponentCanNotBeRemovedException;
 import com.orientechnologies.orient.core.exception.OTooBigIndexKeyException;
 import com.orientechnologies.orient.core.index.OAlwaysGreaterKey;
 import com.orientechnologies.orient.core.index.OAlwaysLessKey;
@@ -486,6 +487,9 @@ public class OSBTree<K, V> extends ODurableComponent {
     try {
       acquireExclusiveLock();
       try {
+        if (size() > 0) {
+          throw new ONonEmptyComponentCanNotBeRemovedException(getName() + " : Only empty tries can be removed");
+        }
         deleteFile(fileId);
 
         if (nullPointerSupport) {
